[feat-82] extracted DND handling function
