+++ src/components/Dashboard/List/List.jsx
@@ -2,7 +2,7 @@ import React, { useContext, useState } from 'react';
 import ViewPositionModal from '../../Modals/ViewPosition/ViewPositionModal';
 import AddPositionModal from '../../Modals/AddPosition/AddPositionModal';
 import Spinner from '../../Spinner/Spinner';
-import Sort from '../Sort/Sort.jsx';
+import onDragEnd from '../../../Utils/Dnd';
 
 import useKanban from '../../../hooks/useKanban';
 import ListHeader from './ListHeader';
@@ -15,8 +15,6 @@ import Messages from '../../Messages/Messages';
 
 import { SelectedPositionContext } from '../../../contexts/SelectedPositionContext';
 import { PositionsContext } from '../../../contexts/PositionsContext';
-
-import { MessagesContext } from '../../../contexts/MessagesContext';
 import { useAuth } from '../../../contexts/AuthContext';
 
 const List = (props = {}) => {
@@ -24,7 +22,6 @@ const List = (props = {}) => {
 
     const { positions } = useContext(PositionsContext);
     const { setSelectedPosition } = useContext(SelectedPositionContext);
-    const { information } = useContext(MessagesContext);
     const { currentUser } = useAuth();
 
     const [isCreateOpen, setIsCreateOpen] = useState(false);
@@ -36,92 +33,14 @@ const List = (props = {}) => {
     };
     const { initialData, setInitialData } = useKanban(currentUser.uid);
 
-    const onDragEnd = (result) => {
-        const { destination, source, draggableId } = result;
-
-        if (!destination) return;
-
-        const startColumn = initialData.columns.find((col) => col.id === source.droppableId);
-        const startColumnIndex = initialData.columns.findIndex((col) => col.id === source.droppableId);
-        const endColumn = initialData.columns.find((col) => col.id === destination.droppableId);
-        const endColumnIndex = initialData.columns.findIndex((col) => col.id === destination.droppableId);
-
-        if (startColumn === endColumn) {
-            const newPositionIds = Array.from(endColumn.positionIds);
-
-            newPositionIds.splice(source.index, 1);
-            newPositionIds.splice(destination.index, 0, draggableId);
-
-            const newColumn = {
-                ...endColumn,
-                positionIds: newPositionIds,
-            };
-
-            const newState = {
-                ...initialData,
-                columns: [...initialData.columns],
-            };
-
-            newState.columns[endColumnIndex] = newColumn;
-
-            app.firestore()
-                .collection('users')
-                .doc(`${currentUser.uid}`)
-                .collection('columns')
-                .doc(`${startColumn.id}`)
-                .update({ positionIds: newPositionIds });
-
-            setInitialData(newState);
-
-            return;
-        }
-
-        const startPositionIDs = Array.from(startColumn.positionIds);
-        startPositionIDs.splice(source.index, 1);
-        const newStart = {
-            ...startColumn,
-            taskIds: startPositionIDs,
-        };
-
-        const finishPositionIDs = Array.from(endColumn.positionIds);
-        finishPositionIDs.splice(destination.index, 0, draggableId);
-        const newFinish = {
-            ...endColumn,
-            taskIds: finishPositionIDs,
-        };
-
-        const newState = {
-            ...initialData,
-            columns: [...initialData.columns],
-        };
-
-        newState.columns[startColumnIndex] = newStart;
-        newState.columns[endColumnIndex] = newFinish;
-
-        app.firestore()
-            .collection('users')
-            .doc(`${currentUser.uid}`)
-            .collection('columns')
-            .doc(`${startColumn.id}`)
-            .update({ positionIds: startPositionIDs });
-
-        app.firestore()
-            .collection('users')
-            .doc(`${currentUser.uid}`)
-            .collection('columns')
-            .doc(`${endColumn.id}`)
-            .update({ positionIds: finishPositionIDs });
-
-        app.firestore().collection('positions').doc(`${draggableId}`).update({ status: endColumn.title });
-        setInitialData(newState);
-    };
+    const endDragHandler = onDragEnd(initialData, currentUser, setInitialData, app);
 
     return (
         <S.ListWrapper>
             {positions.loading ? (
                 <Spinner />
             ) : (
-                <DragDropContext onDragEnd={onDragEnd}>
+                <DragDropContext onDragEnd={endDragHandler}>
                     {initialData &&
                         initialData.columns.map((column, index) => {
                             return (
@@ -139,7 +58,6 @@ const List = (props = {}) => {
                                         uid={currentUser.uid}
                                         setIsViewOpen={setIsViewOpen}
                                         addSelectedPosition={addSelectedPosition}
-                                        positions={initialData.positions}
                                         column={column}
                                         initialData={initialData}
                                     />
