+++ src/components/Dashboard/List/style.js
@@ -9,22 +9,6 @@ export const ListWrapper = styled.div`
     height: inherit;
 `;
 
-export const ListMessages = styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 25px;
-    background: ${({ message }) => (message === 'bad' ? '#fe4c4c' : '#6ACB73')};
-    border-radius: 0px 0px 5px 5px;
-    font-weight: bold;
-    font-size: 14px;
-    line-height: 15px;
-    color: #ffffff;
-    text-align: center;
-    padding: 1px;
-`;
-
 export const List = styled.div`
     display: flex;
     flex-direction: column;
